han-line-bot/requirements.txt
beautifulsoup4==4.5.1
ChatterBot==0.8.0
chatterbot-corpus==1.1.1
click==6.7
Flask==0.12.2
future==0.16.0
itsdangerous==0.24
Jinja2==2.10
jsondatabase==0.1.7
line-bot-sdk==1.5.0
MarkupSafe==1.0
mathparse==0.1.1
nltk==3.2.5
oauthlib==2.0.6
pymongo==3.5.1
python-dateutil==2.6.1
python-twitter==3.3
pytz==2016.7
PyYAML==3.12
requests==2.11.1
requests-oauthlib==0.8.0
six==1.11.0
SQLAlchemy==1.1.15
Werkzeug==0.12.2
